+++ src/index.js
@@ -3,11 +3,8 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
-import { createStore } from 'redux';
 import { Provider } from 'react-redux';
-import { reducer } from './store/reducer';
-
-const store = createStore(reducer);
+import store from './store/configureStore';
 
 ReactDOM.render(
     <Provider store={store} >
