+++ src/store/configureStore.js
@@ -1,0 +1,6 @@
+import rootReducer from './reducers/rootReducer';
+import { createStore } from 'redux';
+
+const store = createStore(rootReducer);
+
+export default store;
