+++ src/store/initialState.js
@@ -1,0 +1,7 @@
+const initialState = {
+    isInArea: false,
+    currentRoom: null,
+    rtc: null
+}
+
+export default initialState;
